Issue #405: Extract user agent methods from WebViewProvider.
They're static so there's no reason they need to be in the provider.
