+++ app/src/amazonWebview/java/org/mozilla/focus/browser/UserAgent.java
@@ -1,0 +1,80 @@
+package org.mozilla.focus.browser;
+
+import android.content.Context;
+import android.content.pm.PackageManager;
+import android.os.Build;
+import android.support.annotation.VisibleForTesting;
+import android.text.TextUtils;
+
+import com.amazon.android.webkit.AmazonWebSettings;
+
+/**
+ * Created by mcomella on 1/24/18.
+ */
+public class UserAgent {
+    /**
+     * Build the browser specific portion of the UA String, based on the webview's existing UA String.
+     */
+    @VisibleForTesting
+    static String getUABrowserString(final String existingUAString, final String focusToken) {
+        // Use the default WebView agent string here for everything after the platform, but insert
+        // Focus in front of Chrome.
+        // E.g. a default webview UA string might be:
+        // Mozilla/5.0 (Linux; Android 7.1.1; Pixel XL Build/NOF26V; wv) AppleWebKit/537.36 (KHTML, like Gecko) Version/4.0 Chrome/56.0.2924.87 Mobile Safari/537.36
+        // And we reuse everything from AppleWebKit onwards, except for adding Focus.
+        int start = existingUAString.indexOf("AppleWebKit");
+        if (start == -1) {
+            // I don't know if any devices don't include AppleWebKit, but given the diversity of Android
+            // devices we should have a fallback: we search for the end of the platform String, and
+            // treat the next token as the start:
+            start = existingUAString.indexOf(")") + 2;
+
+            // If this was located at the very end, then there's nothing we can do, so let's just
+            // return focus:
+            if (start >= existingUAString.length()) {
+                return focusToken;
+            }
+        }
+
+        final String[] tokens = existingUAString.substring(start).split(" ");
+
+        for (int i = 0; i < tokens.length; i++) {
+            if (tokens[i].startsWith("Chrome")) {
+                tokens[i] = focusToken + " " + tokens[i];
+
+                return TextUtils.join(" ", tokens);
+            }
+        }
+
+        // If we didn't find a chrome token, we just append the focus token at the end:
+        return TextUtils.join(" ", tokens) + " " + focusToken;
+    }
+
+    public static String buildUserAgentString(final Context context, final AmazonWebSettings settings, final String appName) {
+        final StringBuilder uaBuilder = new StringBuilder();
+
+        uaBuilder.append("Mozilla/5.0");
+
+        // WebView by default includes "; wv" as part of the platform string, but we're a full browser
+        // so we shouldn't include that.
+        // Most webview based browsers (and chrome), include the device name AND build ID, e.g.
+        // "Pixel XL Build/NOF26V", that seems unnecessary (and not great from a privacy perspective),
+        // so we skip that too.
+        uaBuilder.append(" (Linux; Android ").append(Build.VERSION.RELEASE).append(") ");
+
+        final String existingWebViewUA = settings.getUserAgentString();
+
+        final String appVersion;
+        try {
+            appVersion = context.getPackageManager().getPackageInfo(context.getPackageName(), 0).versionName;
+        } catch (PackageManager.NameNotFoundException e) {
+            // This should be impossible - we should always be able to get information about ourselves:
+            throw new IllegalStateException("Unable find package details for Focus", e);
+        }
+
+        final String focusToken = appName + "/" + appVersion;
+        uaBuilder.append(getUABrowserString(existingWebViewUA, focusToken));
+
+        return uaBuilder.toString();
+    }
+}
