+++ app/src/testAmazonWebview/java/org/mozilla/focus/browser/UserAgentTest.java
@@ -1,4 +1,4 @@
-package org.mozilla.focus.web;
+package org.mozilla.focus.browser;
 
 import android.os.Build;
 import android.webkit.WebSettings;
@@ -7,6 +7,7 @@ import com.amazon.android.webkit.AmazonWebSettings;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
+import org.mozilla.focus.web.WebViewProvider;
 import org.robolectric.RobolectricTestRunner;
 import org.robolectric.RuntimeEnvironment;
 
@@ -15,35 +16,35 @@ import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
 @RunWith(RobolectricTestRunner.class)
-public class WebViewProviderTest {
+public class UserAgentTest {
 
     @Test public void testGetUABrowserString() {
         // Typical situation with a webview UA string from Android 5:
         String focusToken = "Focus/1.0";
         final String existing = "Mozilla/5.0 (Linux; Android 5.0.2; Android SDK built for x86_64 Build/LSY66K) AppleWebKit/537.36 (KHTML, like Gecko) Version/4.0 Chrome/37.0.0.0 Mobile Safari/537.36";
         assertEquals("AppleWebKit/537.36 (KHTML, like Gecko) Version/4.0 " + focusToken + " Chrome/37.0.0.0 Mobile Safari/537.36",
-                WebViewProvider.getUABrowserString(existing, focusToken));
+                UserAgent.getUABrowserString(existing, focusToken));
 
         // Make sure we can use any token, e.g Klar:
         focusToken = "Klar/2.0";
         assertEquals("AppleWebKit/537.36 (KHTML, like Gecko) Version/4.0 " + focusToken + " Chrome/37.0.0.0 Mobile Safari/537.36",
-                WebViewProvider.getUABrowserString(existing, focusToken));
+                UserAgent.getUABrowserString(existing, focusToken));
 
         // And a non-standard UA String, which doesn't contain AppleWebKit
         focusToken = "Focus/1.0";
         final String imaginaryKit = "Mozilla/5.0 (Linux) ImaginaryKit/-10 (KHTML, like Gecko) Version/4.0 Chrome/37.0.0.0 Mobile Safari/537.36";
         assertEquals("ImaginaryKit/-10 (KHTML, like Gecko) Version/4.0 " + focusToken + " Chrome/37.0.0.0 Mobile Safari/537.36",
-                WebViewProvider.getUABrowserString(imaginaryKit, focusToken));
+                UserAgent.getUABrowserString(imaginaryKit, focusToken));
 
         // Another non-standard UA String, this time with no Chrome (in which case we should be appending focus)
         final String chromeless = "Mozilla/5.0 (Linux; Android 5.0.2; Android SDK built for x86_64 Build/LSY66K) AppleWebKit/537.36 (KHTML, like Gecko) Version/4.0 Imaginary/37.0.0.0 Mobile Safari/537.36";
         assertEquals("AppleWebKit/537.36 (KHTML, like Gecko) Version/4.0 Imaginary/37.0.0.0 Mobile Safari/537.36 " + focusToken,
-                WebViewProvider.getUABrowserString(chromeless, focusToken));
+                UserAgent.getUABrowserString(chromeless, focusToken));
 
         // No AppleWebkit, no Chrome
         final String chromelessImaginaryKit = "Mozilla/5.0 (Linux) ImaginaryKit/-10 (KHTML, like Gecko) Version/4.0 Imaginary/37.0.0.0 Mobile Safari/537.36";
         assertEquals("ImaginaryKit/-10 (KHTML, like Gecko) Version/4.0 Imaginary/37.0.0.0 Mobile Safari/537.36 " + focusToken,
-                WebViewProvider.getUABrowserString(chromelessImaginaryKit, focusToken));
+                UserAgent.getUABrowserString(chromelessImaginaryKit, focusToken));
 
     }
 
@@ -58,7 +59,7 @@ public class WebViewProviderTest {
         when(testSettings.getUserAgentString()).thenReturn("Mozilla/5.0 (Linux; U; Android 4.0.3; ko-kr; LG-L160L Build/IML74K) AppleWebkit/534.30 (KHTML, like Gecko) Version/4.0 Mobile Safari/534.30");
 
         assertEquals("Mozilla/5.0 (Linux; Android " + Build.VERSION.RELEASE + ") AppleWebkit/534.30 (KHTML, like Gecko) Version/4.0 Mobile Safari/534.30 fakeappname/null",
-                WebViewProvider.buildUserAgentString(RuntimeEnvironment.application, testSettings, "fakeappname"));
+                UserAgent.buildUserAgentString(RuntimeEnvironment.application, testSettings, "fakeappname"));
     }
 
 }
